Continue to check if a chia pid has disappeared if the log file is not available
Full context is on https://github.com/swar/Swar-Chia-Plot-Manager/issues/246

Due to another bug, the manager has found Chia processes but has not been
able to find an associated log file for that.

The code will currently skip over a pid in this scenario.  This leads to a
situation where Chia processes are finished normally but the manager is
not aware.

This patch does the pid_exists check, even if the logfile is not available.