Update log.py
In windows MadMax stotik's fork logs produces 15 lines of header before Phase 1 starts. This value may be variable (based on how much extra -2 -d -u parameters you've specified), but in the most common scenarios it's around 15. In ideal scenario it should be included in the config progress parameters, so the user itself be able to adjust it for it's personal logs. Currently, original value "madmax=11" produces wrong percentage progress calculations in the viewer.